change type of amount credit

+++ db/migrate/20210723162520_create_credits.rb
@@ -1,7 +1,9 @@
+# frozen_string_literal: true
+
 class CreateCredits < ActiveRecord::Migration[6.1]
   def change
     create_table :credits do |t|
-      t.numeric :amount
+      t.float :amount
 
       t.timestamps
     end
